Work in progress - add PostSharp aspect to record attempts/good/bad counts and timings
git-svn-id: https://scm.dev:8443/svn/lib_statsd/trunk@9 d6dd0074-7a69-5544-9745-f2d122130f32

+++ src/JustEat.Aop/StatsDPipe.cs
@@ -19,12 +19,12 @@ namespace JustEat.Aop
 			_udpClient.Dispose();
 		}
 
-		public bool Timing(string key, int value)
+		public bool Timing(string key, long value)
 		{
 			return Timing(key, value, 1.0);
 		}
 
-		public bool Timing(string key, int value, double sampleRate)
+		public bool Timing(string key, long value, double sampleRate)
 		{
 			return Send(sampleRate, String.Format("{0}:{1:d}|ms", key, value));
 		}
